finish type rename and move migrations
