+++ api/config/custom-environment-variables.json
@@ -1,6 +1,9 @@
 {
-  "postgres": {
-    "url": "POSTGRES_URL"
+  "postgresApi": {
+    "url": "API_POSTGRES_URL"
+   },
+   "postgresGeoApi": {
+    "url": "GEO_POSTGRES_URL"
    },
   "network": {
     "cors": {
