+++ api/src/modules/projects/projects.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { ProjectsController } from './projects.controller';
-import { Project } from './project.entity';
+import { Project } from './project.api.entity';
 import { ProjectsService } from './projects.service';
 
 @Module({
