+++ api/src/modules/users/users.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, Request, UseGuards } from '@nestjs/common';
-import { User } from './user.entity';
+import { User } from './user.api.entity';
 import { UsersService } from './users.service';
 
 import JSONAPISerializer = require('jsonapi-serializer');
