+++ api/src/modules/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { UsersController } from './users.controller';
-import { User } from './user.entity';
+import { User } from './user.api.entity';
 import { UsersService } from './users.service';
 
 @Module({
