+++ geoprocessing/ormconfig.ts
@@ -8,22 +8,32 @@ const config = require('config');
  *
  * ['staging', 'production'].includes(config.util.getEnv('NODE_ENV')) ? true : false
  */
-module.exports = {
+module.exports = [{
+  name: 'default',
   synchronize: false,
   type: 'postgres',
-  url: config.get('postgres.url'),
+  url: config.get('postgresGeoApi.url'),
   ssl: false,
-  entities: ['src/modules/**/*.entity.ts'],
+  entities: ['src/modules/**/*.geo.entity.ts'],
   // Logging may be: ['query', 'error', 'schema', 'warn', 'info', 'log'] Use
   // 'query' if needing to see the actual generated SQL statements (this should
   // be limited to `NODE_ENV=development`). Use 'error' for least verbose
   // logging.
   logging: ['error'],
-  cache: false,
-  migrations: ['src/migrations/**/*.ts'],
-  migrationsRun: true,
-  cli: {
-    migrationsDir: "src/migrations",
-    migrationsTableName: "migrations",
-  }
-};
+  cache: false
+},
+{
+  name: "apiDB",
+  synchronize: false,
+  type: 'postgres',
+  url: config.get('postgresApi.url'),
+  ssl: false,
+  entities: ['src/modules/**/*.api.entity.ts'],
+  // Logging may be: ['query', 'error', 'schema', 'warn', 'info', 'log'] Use
+  // 'query' if needing to see the actual generated SQL statements (this should
+  // be limited to `NODE_ENV=development`). Use 'error' for least verbose
+  // logging.
+  logging: ['error'],
+  cache: false
+},
+];
